refactor: Consolidate ShapeData type definition
Refactors the codebase to have a single source of truth for the `ShapeData` type.

The `ShapeData` interface was previously defined in both `App.tsx` and `src/state/reducer.ts`. This commit removes the duplicate definition from `App.tsx`.

All dependent files (`App.tsx`, `Shape.tsx`, `SvgCanvas.tsx`) now import the type from `src/state/reducer.ts`. This improves maintainability and adheres to the DRY principle.
